Add files via upload
Sign in and sign up work, I have the sign commented out so that we can just go to post screen, the actual firebase accessing is super slow for me, like wait 8 minutes for a response. I started working on create post and post but im not sure if it doesnt work or if its just super slow. Ill hop on it before the weekend if you need help